fix navbar reloading & top bar fontsize
- fix the top bar reloading when viewing single product
- reduce the back icon size and dial message font size
