+++ components/product/single/Topbar/topbar.module.css
@@ -38,3 +38,17 @@
         height:50px;
     }
 }
+
+@media (max-width: 440px){
+    .backIcon img{
+        margin-left: 10px;
+        padding: 5px;
+        width:30px;
+        height:30px;
+    }
+    .code {
+        color: var(--blue-color);
+        font-weight: 400;
+        font-size: 18px;
+    }
+}
